CHomP interface updated to 64-bit

+++ include/database/structures/TreeGrid.h
@@ -164,7 +164,7 @@ public:
   template < class Container >
   size_type getDepth ( const Container & cont ) const;
   
-  void GridElementToCubes (std::vector<std::vector < uint32_t > > * cubes ,
+  void GridElementToCubes (std::vector<std::vector < uint64_t > > * cubes ,
                            const GridElement ge,
                            int depth ) const;
   
@@ -1129,14 +1129,14 @@ TreeGrid::getDepth ( const Container & cont ) const {
 }
 
 inline void 
-TreeGrid::GridElementToCubes ( std::vector<std::vector < uint32_t > > * cubes,
+TreeGrid::GridElementToCubes ( std::vector<std::vector < uint64_t > > * cubes,
                                              const GridElement ge, int depth ) const {
   // Obtain the prefix
   //std::cout << "GEtoCubes: " << geometry ( ge ) << "\n";
   int D = dimension ();
   
   // Determine width
-  typedef std::vector < uint32_t > Cube;
+  typedef std::vector < uint64_t > Cube;
   Cube cube ( D, 0 );
   
   int dim = 0;
@@ -1155,7 +1155,7 @@ TreeGrid::GridElementToCubes ( std::vector<std::vector < uint32_t > > * cubes,
   for ( int d = 0; d < GridElement_depth; ++ d ) {
     if ( dim == D ) dim = 0;
     cube [ dim ] <<= 1;
-    cube [ dim ] |= (uint32_t) p [ p_end - d - 1 ];
+    cube [ dim ] |= (uint64_t) p [ p_end - d - 1 ];
     ++ dim;
   }
   // make the cubes
@@ -1219,7 +1219,7 @@ TreeGrid::relativeComplex ( chomp::RelativePair * pair,
   //std::cout << "XGridElements:\n";
   
   // Make set of cubes, and learn bounds of the cubes.
-  typedef std::vector < uint32_t > Cube;
+  typedef std::vector < uint64_t > Cube;
   Cube mincube ( D, -1 );
   Cube maxcube ( D, 0 );
   RectGeo newbounds ( D );
@@ -1258,7 +1258,7 @@ int percent = 0;
     }
   }
   
-  std::vector < uint32_t > dimension_sizes ( D, 1 );
+  std::vector < uint64_t > dimension_sizes ( D, 1 );
   std::vector < bool > is_periodic = periodic_;
   for ( int d = 0; d < D; ++ d ) {
     dimension_sizes [ d ] = maxcube [ d ] - mincube [ d ] + 1;
@@ -1274,7 +1274,7 @@ int percent = 0;
   BOOST_FOREACH ( GridElement e, XGridElements ) {
     //std::cout << e << "\n";
     //std::cout << "GEOMETRY = " << geometry ( e ) << "\n";
-    typedef std::vector < uint32_t > Cube;
+    typedef std::vector < uint64_t > Cube;
     std::vector < Cube > cubes;
     GridElementToCubes ( &cubes, e, depth );
     BOOST_FOREACH ( Cube & cube, cubes ) {
@@ -1297,7 +1297,7 @@ int percent = 0;
   //std::cout << "AGridElements:\n";
   BOOST_FOREACH ( GridElement e, AGridElements ) {
     //std::cout << e << "\n";
-    typedef std::vector < uint32_t > Cube;
+    typedef std::vector < uint64_t > Cube;
     std::vector < Cube > cubes;
     GridElementToCubes ( &cubes, e, depth );
     BOOST_FOREACH ( Cube & cube, cubes ) {
